6. MCP/3. mcp client/requirements.txt
python-dotenv # 환경변수를 파일(.env)로 관리하는 라이브러리
langchain_mcp_adapters
langgraph
langchain_openai